chapter7_scratch/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
anyio==3.7.1
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
boto3==1.26.122
botocore==1.29.122
Brotli==1.1.0
celery==5.3.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.6.3
dnspython==2.4.2
duckduckgo-search==3.9.11
Flask==2.3.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.4
langchain==0.0.347
langchain-core==0.0.12
langfuse==1.0.18
langsmith==0.0.69
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==0.27.8
packaging==23.2
pinecone-client==2.2.2
prompt-toolkit==3.0.41
pydantic==1.10.13
pypdf==3.15.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
regex==2023.10.3
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uuid==1.30
vine==5.1.0
watchdog==3.0.0
wcwidth==0.2.12
Werkzeug==2.3.1
yarl==1.9.4
